Early sketch of the parser, minimize `poll` abstraction
This is mainly committed now so that I can also commit the change to
remove all the extra machinery around `PollSet` and switch to a (rather
large) function.

`Read` from the read FD needs a mutable reference so we needed to make
some changes like returning `[bool; 3]` when introducing the parser.
Parser is very early though and that code is a dummy for now.