Allow polling and reading from the Terminal instance
This matches crossterm's global `poll` and `read` functions. To make
this possible we need store the `InternalEventReader` on the terminal.
In order to make this play nicely with the `EventStream` we need to
expose the `waker` on `InternalEventReader` and have `EventStream::new`
take a clone of the reader, using that new function to get the waker.