Bound Terminal trait on io::Write
Termwiz doesn't do this because it has a change API where it tries to
optimize operations. Instead what I'd like in Termina is something
closer to Crossterm's `queue!` and `execute!` macros, which can be
emulated now with `write!` and `terminal.flush()`. In fact that's what
is going on under the hook in those macros, it's just that they write to
stdout or the tty.