Merge InternalEvent and Event together
It would be useful as a dependent of this library to be able to read
certain events and leave the rest for the event stream. This change
merges the `InternalEvent` type into `Event` (i.e. by adding a Csi
variant to `Event`).

To accommodate this, the `event_stream`, `poll` and `read` members from
the `Terminal` trait now accept a filter function which is passed
through to `InternalEventReader` (now just `EventReader`).

A use-case for the filters can be seen in the `event-read` example. In
the function which merges resizes events together
(`flush_resize_events`), we now poll and read specifically resize
events. Previously it was possible to lose other events while flushing
resizes. It was unlikely during normal use, since you don't usually
resize the window while doing some other kind of input, but possible in
applications like Kitty which have a resizing widget for panes: focus
events could be lost while resizing in the example. Now that we can
poll specifically for window resizes this is not an issue.