Add types for merging multiple SGR escapes into one
This new `SgrAttributes` type can contain all of the common `Sgr`
attributes together. It's like a `Vec<Sgr>` but flattened into a struct.
When displayed, all SGR updates are made at once. See the example in
the docs for the type: we can emit a change to set the foreground as
green and bold as 'CSI 32 ; 1 m' instead of two separate 'CSI 32 m' plus
'CSI 1 m' escapes. This is equivalent when writing exactly one SGR but
is a savings of two bytes (add one for ';', subtract three for CSI and
'm') per extra change made in the SGR.

A similar change in crossterm reportedly boosted FPS of an application
that sets a combined foreground+backgrouond SGR for every cell by
15-25%: <https://redirect.github.com/crossterm-rs/crossterm/pull/879>.
This type takes that change a step further and supports all SGR
attributes.