Remove `Terminal::reset_mode`
This was a footgun since it isn't an opposite to `enter_raw_mode`. We
want to reset the mode in Helix when suspending but that only works on
Unix and on Unix it's the same thing to use `enter_cooked_mode`. On
Windows during Drop we want to reset the code pages (encoding) as well
but the code pages are not set in `enter_raw_mode`, they're set during
`WindowsTerminal::new`. This change removes `reset_mode` and moves its
code into `Drop for WindowsTerminal`.