Chunk CSI attributes to fit conservative Ps limits
See <https://redirect.github.com/helix-editor/helix/pull/13307#issuecomment-2813435388>.
Some terminals' VT parsers have conservative limits for the number of
parameters they will accept, sometimes as low as 10. For compatibility
with these terminals we should chunk the combined SGR attributes to fit
that limit. This is added as an option on `SgrAttributes` so that it can
be customized and changed dynamically. (As opposed to a constant.)