Expose the EventReader, make EventStream optional
We can gate EventStream behind a feature flag to make the futures-core
dependency optional. You may wish to roll your own event stream to avoid
futures and async. This change replaces `Terminal::event_stream` with
`Terminal::event_reader` and makes the `EventReader` type `pub` instead
of `pub(crate)`.

This change also removes the `DummyEventSource` since it is easy to
write that type in the consuming application.