Windows: Fix zero/one-based handling of window resize events
There was a mistake with the WindowResized event on Windows from some
code inherited from crossterm. Crossterm adds one to the dimensions
returned from a window resize event because it overwrites the dimensions
in the event with ones from `GetConsoleScreenBufferInfo` (zero-based).
The rows and columns in the event are already one-based.

This change also re-types the dimensions with the `OneBased` helper from
termwiz. The `OneBased` type is moved from the `escape` module to the
crate root.

Fixes #5.